requirements.txt
asgiref==3.2.3
Django==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
uuid==1.30
